third_party/WebKit/LayoutTests/http/tests/inspector/network/network-blocked-reason.html

<html>
<head>
<meta http-equiv="Content-Security-Policy" content="script-src 'self' 'unsafe-inline'">
<script src="/inspector/inspector-test.js"></script>
<script src="/inspector/network-test.js"></script>
<script>

function sendCSPRequest()
{
    var script = document.createElement("script");
    script.src = "https://www.example.com/csp.js";
    document.head.appendChild(script);
}

function addBlockedScript(url)
{
    var script = document.createElement("script");
    script.src = url;
    document.head.appendChild(script);
}

async function test()
{
    var pane = new Network.BlockedURLsPane();
    pane._requestBlockingEnabled = true;
    await pane._update();

    var waitForRequest = InspectorTest.waitForEvent.bind(InspectorTest, SDK.NetworkManager.Events.RequestFinished, InspectorTest.networkManager);

    var url = "csp.js";
    InspectorTest.evaluateInPage("sendCSPRequest()");
    var request = await waitForRequest(request => request.name() === url);
    dumpRequestInfo(request);

    await testBlockedURL(["*resources**/silent*.js"], "resources/silent_script.js");
    await testBlockedURL(["a*b"], "ba");
    await testBlockedURL(["***pattern***"], "there/is/a/pattern/inside.js");
    await testBlockedURL(["*pattern*"], "patt1ern");
    await testBlockedURL(["*this***is*a*pattern"], "file/this/is/the/pattern");
    await testBlockedURL(["*this***is*a*pattern"], "this/is/a/pattern");
    await testBlockedURL(["*this***is*a*pattern"], "this/is");

    await testBlockedURL(["pattern"], "long/pattern/inside");
    await testBlockedURL(["*pattern"], "long/pattern/inside");
    await testBlockedURL(["*pattern*"], "long/pattern/inside");
    await testBlockedURL(["pattern*"], "long/pattern/inside");
    await testBlockedURL(["*pattern"], "long/pattern");
    await testBlockedURL(["*pattern*"], "long/pattern");
    await testBlockedURL(["pattern"], "pattern/right");
    await testBlockedURL(["*pattern"], "pattern/right");
    await testBlockedURL(["*pattern*"], "pattern/right");
    await testBlockedURL(["pattern*"], "pattern/right");
    await testBlockedURL(["pattern"], "pattern");
    await testBlockedURL(["*pattern"], "pattern");
    await testBlockedURL(["*pattern*"], "pattern");
    await testBlockedURL(["pattern*"], "pattern");

    await testBlockedURL(["*pattern*", "*pattern*"], "pattern");
    await testBlockedURL(["*a*b*c*d*e*"], "edcbaedcbaedcbaedcba");
    await testBlockedURL(["*a*b*c*d*e*"], "edcbaedcbaedcbaedcbae");
    await testBlockedURL(["*one1*", "*two2"], "one1two2");
    await testBlockedURL(["*one1*", "*two2*", "three3"], "four4");
    await testBlockedURL(["*one1*", "*two2*", "*three3*"], "only-two2-here");
    await testBlockedURL([], "resources/silent_script.js");

    InspectorTest.completeTest()

    async function testBlockedURL(patterns, url) {
        InspectorTest.addResult("Blocked patterns: " + patterns.join(";"));
        InspectorTest.addResult("Request: [DOMAIN_DATA...]" + url);
        pane._patternEntries = patterns.map(pattern => ({pattern: pattern, blockedCount: 0, enabled: true}));
        await pane._update();
        await InspectorTest.clearNetworkCache();

        var requestName = url.substring(url.lastIndexOf("/") + 1);
        InspectorTest.evaluateInPage("addBlockedScript(\"" + url + "\")");
        var request = await waitForRequest(request => request.name() === requestName);
        dumpRequestInfo(request);
    }

    function dumpRequestInfo(request) {
        InspectorTest.addResult("statusCode: " + request.statusCode);
        InspectorTest.addResult("BlockedReason: " + request.blockedReason());
        InspectorTest.addResult("Failed: " + !!request.failed);
        InspectorTest.addResult("Canceled: " + !!request.canceled);
        InspectorTest.addResult("LocalizedFailDescription: " + request.localizedFailDescription);
        InspectorTest.addResult("");
    }
}
</script>
</head>
<body onload="runTest()">
<p>Tests that blocked reason is recognized correctly.</p>
</body>
</html>
