third_party/WebKit/LayoutTests/http/tests/inspector/network/network-blocked-reason-expected.txt

CONSOLE ERROR: line 12: Refused to load the script 'https://www.example.com/csp.js' because it violates the following Content Security Policy directive: "script-src 'self' 'unsafe-inline'".

Tests that blocked reason is recognized correctly.

statusCode: 0
BlockedReason: csp
Failed: true
Canceled: false
LocalizedFailDescription: 

Blocked patterns: *resources**/silent*.js
Request: [DOMAIN_DATA...]resources/silent_script.js
statusCode: 0
BlockedReason: undefined
Failed: true
Canceled: false
LocalizedFailDescription: net::ERR_BLOCKED_BY_CLIENT

Blocked patterns: a*b
Request: [DOMAIN_DATA...]ba
statusCode: 404
BlockedReason: undefined
Failed: true
Canceled: true
LocalizedFailDescription: net::ERR_ABORTED

Blocked patterns: ***pattern***
Request: [DOMAIN_DATA...]there/is/a/pattern/inside.js
statusCode: 0
BlockedReason: undefined
Failed: true
Canceled: false
LocalizedFailDescription: net::ERR_BLOCKED_BY_CLIENT

Blocked patterns: *pattern*
Request: [DOMAIN_DATA...]patt1ern
statusCode: 404
BlockedReason: undefined
Failed: true
Canceled: true
LocalizedFailDescription: net::ERR_ABORTED

Blocked patterns: *this***is*a*pattern
Request: [DOMAIN_DATA...]file/this/is/the/pattern
statusCode: 404
BlockedReason: undefined
Failed: true
Canceled: true
LocalizedFailDescription: net::ERR_ABORTED

Blocked patterns: *this***is*a*pattern
Request: [DOMAIN_DATA...]this/is/a/pattern
statusCode: 0
BlockedReason: undefined
Failed: true
Canceled: false
LocalizedFailDescription: net::ERR_BLOCKED_BY_CLIENT

Blocked patterns: *this***is*a*pattern
Request: [DOMAIN_DATA...]this/is
statusCode: 404
BlockedReason: undefined
Failed: true
Canceled: true
LocalizedFailDescription: net::ERR_ABORTED

Blocked patterns: pattern
Request: [DOMAIN_DATA...]long/pattern/inside
statusCode: 404
BlockedReason: undefined
Failed: true
Canceled: true
LocalizedFailDescription: net::ERR_ABORTED

Blocked patterns: *pattern
Request: [DOMAIN_DATA...]long/pattern/inside
statusCode: 404
BlockedReason: undefined
Failed: true
Canceled: true
LocalizedFailDescription: net::ERR_ABORTED

Blocked patterns: *pattern*
Request: [DOMAIN_DATA...]long/pattern/inside
statusCode: 0
BlockedReason: undefined
Failed: true
Canceled: false
LocalizedFailDescription: net::ERR_BLOCKED_BY_CLIENT

Blocked patterns: pattern*
Request: [DOMAIN_DATA...]long/pattern/inside
statusCode: 404
BlockedReason: undefined
Failed: true
Canceled: true
LocalizedFailDescription: net::ERR_ABORTED

Blocked patterns: *pattern
Request: [DOMAIN_DATA...]long/pattern
statusCode: 0
BlockedReason: undefined
Failed: true
Canceled: false
LocalizedFailDescription: net::ERR_BLOCKED_BY_CLIENT

Blocked patterns: *pattern*
Request: [DOMAIN_DATA...]long/pattern
statusCode: 0
BlockedReason: undefined
Failed: true
Canceled: false
LocalizedFailDescription: net::ERR_BLOCKED_BY_CLIENT

Blocked patterns: pattern
Request: [DOMAIN_DATA...]pattern/right
statusCode: 404
BlockedReason: undefined
Failed: true
Canceled: true
LocalizedFailDescription: net::ERR_ABORTED

Blocked patterns: *pattern
Request: [DOMAIN_DATA...]pattern/right
statusCode: 404
BlockedReason: undefined
Failed: true
Canceled: true
LocalizedFailDescription: net::ERR_ABORTED

Blocked patterns: *pattern*
Request: [DOMAIN_DATA...]pattern/right
statusCode: 0
BlockedReason: undefined
Failed: true
Canceled: false
LocalizedFailDescription: net::ERR_BLOCKED_BY_CLIENT

Blocked patterns: pattern*
Request: [DOMAIN_DATA...]pattern/right
statusCode: 404
BlockedReason: undefined
Failed: true
Canceled: true
LocalizedFailDescription: net::ERR_ABORTED

Blocked patterns: pattern
Request: [DOMAIN_DATA...]pattern
statusCode: 404
BlockedReason: undefined
Failed: true
Canceled: true
LocalizedFailDescription: net::ERR_ABORTED

Blocked patterns: *pattern
Request: [DOMAIN_DATA...]pattern
statusCode: 0
BlockedReason: undefined
Failed: true
Canceled: false
LocalizedFailDescription: net::ERR_BLOCKED_BY_CLIENT

Blocked patterns: *pattern*
Request: [DOMAIN_DATA...]pattern
statusCode: 0
BlockedReason: undefined
Failed: true
Canceled: false
LocalizedFailDescription: net::ERR_BLOCKED_BY_CLIENT

Blocked patterns: pattern*
Request: [DOMAIN_DATA...]pattern
statusCode: 404
BlockedReason: undefined
Failed: true
Canceled: true
LocalizedFailDescription: net::ERR_ABORTED

Blocked patterns: *pattern*;*pattern*
Request: [DOMAIN_DATA...]pattern
statusCode: 0
BlockedReason: undefined
Failed: true
Canceled: false
LocalizedFailDescription: net::ERR_BLOCKED_BY_CLIENT

Blocked patterns: *a*b*c*d*e*
Request: [DOMAIN_DATA...]edcbaedcbaedcbaedcba
statusCode: 404
BlockedReason: undefined
Failed: true
Canceled: true
LocalizedFailDescription: net::ERR_ABORTED

Blocked patterns: *a*b*c*d*e*
Request: [DOMAIN_DATA...]edcbaedcbaedcbaedcbae
statusCode: 0
BlockedReason: undefined
Failed: true
Canceled: false
LocalizedFailDescription: net::ERR_BLOCKED_BY_CLIENT

Blocked patterns: *one1*;*two2
Request: [DOMAIN_DATA...]one1two2
statusCode: 0
BlockedReason: undefined
Failed: true
Canceled: false
LocalizedFailDescription: net::ERR_BLOCKED_BY_CLIENT

Blocked patterns: *one1*;*two2*;three3
Request: [DOMAIN_DATA...]four4
statusCode: 404
BlockedReason: undefined
Failed: true
Canceled: true
LocalizedFailDescription: net::ERR_ABORTED

Blocked patterns: *one1*;*two2*;*three3*
Request: [DOMAIN_DATA...]only-two2-here
statusCode: 0
BlockedReason: undefined
Failed: true
Canceled: false
LocalizedFailDescription: net::ERR_BLOCKED_BY_CLIENT

Blocked patterns: 
Request: [DOMAIN_DATA...]resources/silent_script.js
statusCode: 200
BlockedReason: undefined
Failed: false
Canceled: false
LocalizedFailDescription: undefined
